agent/pyproject.toml
[project]
name = "monet-agent"
version = "0.1.0"
description = "Monet Agent — Autonomous AI stock trading agent with persistent memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.11",
    "langchain-anthropic>=1.3.2",
    "langgraph-cli[inmem]>=0.4.12",
    "langgraph-sdk>=0.1.0",
    "httpx>=0.27.0",
    "supabase>=2.0.0",
    "tavily-python>=0.7.21",
    "alpaca-py>=0.33.0",
    "yfinance>=0.2.40",
    "finnhub-python>=2.4.0",
    "lxml>=5.0.0",
    "ta>=0.11.0",
    "pandas>=2.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/stock_agent"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
